+++ scripts/BUILD.bazel
@@ -15,32 +15,17 @@ kt_jvm_binary(
     name = "pattern_validation_check",
     testonly = True,
     data = REGEX_PATTERN_CHECK_ASSETS,
-    main_class = "org.oppia.android.scripts.RegexPatternValidationCheck$Companion",
-    runtime_deps = [":pattern_validation_lib"],
-)
-
-kt_jvm_library(
-    name = "pattern_validation_lib",
-    testonly = True,
-    srcs = [
-        "src/java/org/oppia/android/scripts/RegexPatternValidationCheck.kt",
-        "src/java/org/oppia/android/scripts/RepoFile.kt",
-        "src/java/org/oppia/android/scripts/ScriptResultConstants.kt",
-    ],
-    deps = [
-        "//scripts/src/java/org/oppia/android/scripts/proto:file_content_validation_structure_java_proto_lite",
-        "//scripts/src/java/org/oppia/android/scripts/proto:filename_pattern_validation_structure_java_proto_lite",
-    ],
+    main_class = "org.oppia.android.scripts.regex.RegexPatternValidationCheckKt",
+    runtime_deps = ["//scripts/src/java/org/oppia/android/scripts/regex:pattern_validation_lib"],
 )
 
 kt_jvm_test(
-    name = "test_pattern_validation_check",
+    name = "RegexPatternValidationCheckTest",
     testonly = True,
-    srcs = ["src/test/org/oppia/android/scripts/RegexPatternValidationCheckTest.kt"],
+    srcs = ["src/test/org/oppia/android/scripts/regex/RegexPatternValidationCheckTest.kt"],
     data = REGEX_PATTERN_CHECK_ASSETS,
-    test_class = "org.oppia.android.scripts.RegexPatternValidationCheckTest",
     deps = [
-        ":pattern_validation_lib",
+        "//scripts/src/java/org/oppia/android/scripts/regex:pattern_validation_lib",
         "//testing:assertion_helpers",
         "//third_party:com_google_truth_truth",
         "//third_party:org_jetbrains_kotlin_kotlin-test-junit",
