+++ scripts/assets/filename_pattern_validation_checks.textproto
@@ -1,4 +1,4 @@
 filename_checks {
-  prohibited_filename_regex: "data/src/main/.+?Activity.kt"
-  failure_message: "Activities cannot be placed in the data module"
+  prohibited_filename_regex: "^((?!(app|testing)).)+/src/main/.+?Activity.kt"
+  failure_message: "Activities cannot be placed outside the app or testing module"
 }
