+++ scripts/src/java/org/oppia/android/scripts/proto/BUILD.bazel
@@ -9,7 +9,7 @@ proto_library(
 
 java_lite_proto_library(
     name = "filename_pattern_validation_structure_java_proto_lite",
-    visibility = ["//scripts:__pkg__"],
+    visibility = ["//scripts/src/java/org/oppia/android/scripts/regex:__pkg__"],
     deps = [":filename_pattern_validation_structure_proto"],
 )
 
@@ -21,6 +21,6 @@ proto_library(
 
 java_lite_proto_library(
     name = "file_content_validation_structure_java_proto_lite",
-    visibility = ["//scripts:__pkg__"],
+    visibility = ["//scripts/src/java/org/oppia/android/scripts/regex:__pkg__"],
     deps = [":file_content_validation_structure_proto"],
 )
